+++ src/menu_bagl.c
@@ -7,15 +7,6 @@
 
 static char blindsign_text[20];
 
-bool is_blindsign_enabled(void) {
-    return N_storage.blindsign_enabled == 0x01;
-}
-
-void toggle_blindsign(void) {
-    uint8_t value = N_storage.blindsign_enabled ? 0x00 : 0x01;
-    nvm_write((void*)&N_storage.blindsign_enabled, &value, sizeof(value));
-}
-
 static void update_blindsign_text(void) {
     if (is_blindsign_enabled()) {
         strcpy(blindsign_text, "Enabled");
